requirements.txt
Django==1.7.1
cloudinary==1.0.18
dj-database-url==0.3.0
dj-static==0.0.6
django-registration==1.0
django-taggit==0.12.2
django-toolbelt==0.0.1
gunicorn==19.1.1
psycopg2==2.5.4
six==1.8.0
static3==0.5.1
